byte to int data type

+++ HunterPie.Integrations/Datasources/MonsterHunterRise/Definitions/MHREquipmentSkillStructure.cs
@@ -5,6 +5,6 @@ namespace HunterPie.Integrations.Datasources.MonsterHunterRise.Definitions;
 [StructLayout(LayoutKind.Explicit)]
 public struct MHREquipmentSkillStructure
 {
-    [FieldOffset(0x10)] public byte Id;
+    [FieldOffset(0x10)] public int Id;
     [FieldOffset(0x14)] public uint Level;
 }
